+++ tests/testthat/test-slice.R
@@ -194,7 +194,7 @@ test_that("can `vec_slice()` records", {
   out <- vec_slice(new_rcrd(list(a = 1L, b = 2L)), rep(1, 3))
   expect_size(out, 3)
 
-  out <- vec_na(new_rcrd(list(a = 1L, b = 2L)), 2)
+  out <- vec_init(new_rcrd(list(a = 1L, b = 2L)), 2)
   expect_size(out, 2)
 })
 
@@ -372,49 +372,49 @@ test_that("missing indices don't create NA names", {
 })
 
 
-# vec_na ------------------------------------------------------------------
+# vec_init ----------------------------------------------------------------
 
 test_that("vec_slice throws error with non-vector inputs", {
   expect_error(vec_slice(environment(), 1L), class = "vctrs_error_scalar_type")
 })
 
-# vec_na ------------------------------------------------------------------
+# vec_init ----------------------------------------------------------------
 
 test_that("na of atomic vectors is as expected", {
-  expect_equal(vec_na(TRUE), NA)
-  expect_equal(vec_na(1L), NA_integer_)
-  expect_equal(vec_na(1), NA_real_)
-  expect_equal(vec_na("x"), NA_character_)
-  expect_equal(vec_na(1i), NA_complex_)
+  expect_equal(vec_init(TRUE), NA)
+  expect_equal(vec_init(1L), NA_integer_)
+  expect_equal(vec_init(1), NA_real_)
+  expect_equal(vec_init("x"), NA_character_)
+  expect_equal(vec_init(1i), NA_complex_)
 })
 
 test_that("na of factor preserves levels", {
   f1 <- factor("a", levels = c("a", "b"))
-  f2 <- vec_na(f1)
+  f2 <- vec_init(f1)
 
   expect_equal(levels(f1), levels(f2))
 })
 
 test_that("na of POSIXct preserves tz", {
   dt1 <- as.POSIXct("2010-01-01", tz = "America/New_York")
-  dt2 <- vec_na(dt1)
+  dt2 <- vec_init(dt1)
   expect_equal(attr(dt2, "tzone"), "America/New_York")
 })
 
 test_that("na of list is list(NULL)", {
-  expect_equal(vec_na(list()), list(NULL))
+  expect_equal(vec_init(list()), list(NULL))
 })
 
 test_that("na of array is 1d slice", {
   x1 <- array(1:12, c(2, 3, 4))
-  x2 <- vec_na(x1)
+  x2 <- vec_init(x1)
 
   expect_equal(x2, array(NA_integer_, c(1, 3, 4)))
 })
 
 test_that("na of list-array is 1d slice", {
   x1 <- array(as.list(1:12), c(2, 3, 4))
-  x2 <- vec_na(x1)
+  x2 <- vec_init(x1)
 
   expect_equal(x2, array(list(), c(1, 3, 4)))
 })